Added The Journey World
The Journey, an interactive experience of self-discovery, made during the Winter 2025 Term of Design Studio 3 by Cole McMullin, Alex Di Nobile, Daniel Gagné, Cohen Ly and Conall Nelligan. The Source Code for circles has been edited in the circles-user-networked.js and circles-portal.js scripts, and may need to be adapted to  not break the experiences for other worlds.